moved CreateModelOptions to Model namespace, changed using statement for PersonalityInsights from Models to Model - I chose wrong in the diff
